requirements.txt
alembic==1.4.2
attrs==19.3.0
azure-cognitiveservices-vision-computervision==0.6.0
azure-common==1.1.25
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cheroot==8.3.0
click==7.1.2
cycler==0.10.0
flasgger==0.9.4
Flask==1.1.2
Flask-Admin==1.5.6
Flask-Login==0.5.0
Flask-Material==0.1.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
flaskweb==0.2.1
gevent==20.6.2
greenlet==0.4.16
gunicorn==20.0.4
http-client==0.1.22
idna==2.9
importlib-metadata==1.6.1
isodate==0.6.0
itsdangerous==1.1.0
jaraco.functools==3.0.1
Jinja2==2.11.2
jsonschema==3.2.0
kiwisolver==1.2.0
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.2.2
mistune==0.8.4
more-itertools==8.4.0
msrest==0.6.16
numpy==1.19.0
oauthlib==3.1.0
pandas==1.0.5
Pillow==7.1.2
pycparser==2.20
pycurl==7.43.0.5
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
SQLAlchemy==1.3.17
urllib3==1.25.9
WTForms==2.3.1
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0